Bug 124 - ShowMonthPicker, MonthLabelColor, and DaysTitleColor values are not honored (#125)
Fixed the references to ShowLayoutUnitPicker and LayoutUnitLabelColor to reference ShowMonthPicker and MonthLabelColor  respectfully.

Changed IViewLayoutEngine.GenerateLayout to take in the bindings for the calendar properties so the individual controls can be bound to the referenced properties, so any runtime changes to those values will automatically be updated in the contained controls. This also fixed the DaysTitleColor issue.